requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
gspread==5.7.0
oauth2client==4.1.3
slack_sdk==3.17.0
